Updated description and fixed visualization bug
